test: shared integration-test helper extensions + named timeouts (#62)
